Updated to not write transactions when server errors.
